# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Farmacas",
        "version": "0.1"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authenticate"
                ],
                "summary": "Realiza o login do usuário",
                "description": "Este endpoint permite que um usuário faça o login fornecendo um email e uma senha válidos.",
                "operationId": "8dcb70df1020986038d098cc08d05dae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Email do usuário",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "description": "Senha do usuário",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login bem-sucedido. Retorna um token de autenticação.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "description": "Token de acesso",
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "description": "Tipo de token (Bearer)",
                                            "type": "string"
                                        },
                                        "expires_in": {
                                            "description": "Tempo de expiração do token em segundos",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas. Falha na autenticação."
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Authenticate"
                ],
                "summary": "Registra um novo usuário",
                "description": "Este endpoint permite que um novo usuário seja registrado fornecendo informações necessárias, como nome, email e senha.",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Nome do usuário",
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "email": {
                                        "description": "Email do usuário",
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 255
                                    },
                                    "password": {
                                        "description": "Senha do usuário",
                                        "type": "string",
                                        "maxLength": 255,
                                        "minLength": 6
                                    },
                                    "password_confirmation": {
                                        "description": "Confirmação da senha do usuário",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário registrado com sucesso. Retorna um token de autenticação.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "description": "Token de acesso",
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "description": "Tipo de token (Bearer)",
                                            "type": "string"
                                        },
                                        "expires_in": {
                                            "description": "Tempo de expiração do token em segundos",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação. Verifique os campos fornecidos."
                    }
                }
            }
        },
...
    },
...
}